[SYNCE Plugin] Fixed a settings update bug in the synce plugin.
The Synce plugin maintains its own cache of synchronisation settngs
and this cache was not updated when an account was added, deleted,
or modified.  This caused incorrect settings to be returned if
two device management session were run one after another in the
same provman-session process instance.  The values reported by
the synce plugin during the second session were incorrect, as
they had not been updated to reflect any changes made during
the first.