[EDS Plugin] Fixed a small bug when adding new accounts
When adding and updating email accounts the eds plugin maintains
a hash table of accounts that are being updated.  The key of this
hash table is the plugin id.  However, when new accounts were being
added the local key was being used.  This didn't create any
noticeable problems but it was inefficient so the bug has been fixed.