[oFono Plugin] Fixed a settings update bug in the ofono plugin.
The oFono plugin maintains its own cache of telephony settngs
and this cache was not updated when a context was deleted.
This caused incorrect settings to be returned if two device
management session were run one after another in the same
provman-system process instance.  The values reported by the
oFono plugin during the second session were incorrect, as
they had not been updated to reflect any changes made during
the first.  The oFono plugin now discards its entire cache
at the end of a session and reloads the cache when a new session
begins, to ensure that its cache is always up to date.