[PROVMAN] Implemented the Provman cache.
During a session provman now creates a tree structure
called the cache to store all the settings.  Previously
this information was maintained in a hash table of settings
and the existence of directories needed to be inferred.  This
complicated the implementation of certain functions such as
Delete and GetAll.  It also was incovenient for implementing
meta data as we need to know definitively whether or not
a directory exists before we can associate meta data with it.