[PROVMAN] Added the com.intel.provman.Error.Died d-Bus error.
This is generated when a task has been queued for execution but
provman is killed before it gets around to executing the task.
The difference between Died and Cancelled is that the Cancelled
error message is generated for a task that provman is actually
in the process of executing when it is killed.