[PROVMAN] Meta data is no longer lost when a setting's value is overwriten.
There was a bug which caused a setting's meta data to be lost if the value
of the setting was changed.  The set command always created a new node
in the cache.  Now it checks to see if the node already exists.  If it does
it simply overwrites the existing node's value.  If it does not it creates
a new node as before.